Stream hub community links via use() hook
Move the /hub links fetch off the blocking await path: page.tsx now
passes the fetch promise straight to a client component that unwraps
it with use(), wrapped in Suspense, matching the pattern already used
for the community list on the homepage.
